requirements.txt
# Multi-Agent Grounded RAG System — Dependencies
# Python 3.10+

# Core ML / NLP
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0

# Retrieval
faiss-cpu>=1.7.4
chromadb>=0.4.0
rank_bm25>=0.2.2

# Document ingestion
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
pypdf>=3.0.0

# API & server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Orchestration
langgraph>=0.0.20

# Observability
langfuse>=2.0.0

# Evaluation
ragas>=0.1.0
datasets>=2.14.0
matplotlib>=3.5.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.65.0
